backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings

# Database - Using Neon DB (PostgreSQL)
psycopg2-binary>=2.9.0
redis>=5.0.0

# AI/ML Integration
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# File Storage & Processing
boto3
pillow
pypdf2
python-magic
aiofiles

# Background Tasks
celery
apscheduler

# Development
pytest
pytest-asyncio
black
isort
flake8

# Environment
python-dotenv
